+++ refinedstorage-fabric/src/main/java/com/refinedmods/refinedstorage/fabric/PlatformImpl.java
@@ -472,6 +472,25 @@ public final class PlatformImpl extends AbstractPlatform {
         }
     }
 
+    @Override
+    public void requestModelDataUpdateOnClient(final BlockEntity blockEntity, final boolean updateChunk) {
+        final Level level = blockEntity.getLevel();
+        if (level == null) {
+            return;
+        }
+        if (!level.isClientSide()) {
+            throw new IllegalArgumentException("Cannot request model data update on server");
+        }
+        if (updateChunk) {
+            level.sendBlockUpdated(
+                blockEntity.getBlockPos(),
+                blockEntity.getBlockState(),
+                blockEntity.getBlockState(),
+                Block.UPDATE_ALL
+            );
+        }
+    }
+
     private void doSave(final CompoundTag compoundTag, final Path tempFile, final Path targetPath) throws IOException {
         // Write to temp file first.
         NbtIo.writeCompressed(compoundTag, tempFile);