+++ refinedstorage-neoforge/src/main/java/com/refinedmods/refinedstorage/neoforge/autocrafting/FluidHandlerExternalPatternProviderSink.java
@@ -3,6 +3,7 @@ package com.refinedmods.refinedstorage.neoforge.autocrafting;
 import com.refinedmods.refinedstorage.api.autocrafting.task.ExternalPatternSink;
 import com.refinedmods.refinedstorage.api.core.Action;
 import com.refinedmods.refinedstorage.api.resource.ResourceAmount;
+import com.refinedmods.refinedstorage.api.resource.ResourceKey;
 import com.refinedmods.refinedstorage.common.api.autocrafting.PlatformPatternProviderExternalPatternSink;
 import com.refinedmods.refinedstorage.common.support.resource.FluidResource;
 import com.refinedmods.refinedstorage.neoforge.storage.CapabilityCache;
@@ -76,5 +77,10 @@ class FluidHandlerExternalPatternProviderSink implements PlatformPatternProvider
             }
             return true;
         }).orElse(true);
+    }
+
+    @Override
+    public boolean applies(final ResourceKey resource) {
+        return resource instanceof FluidResource;
     }
 }