+++ CHANGELOG.md
@@ -10,6 +10,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ### Fixed
 
 -   Cables being disconnected visually when loading a world.
+-   Fixed processing patterns inserting ingredient types in the wrong order into machines.
 
 ## [2.0.0-beta.15] - 2025-09-24